refactor(build): some random dev build changes
