+++ _scripts/install_all.sh
@@ -1,3 +1,8 @@
 #!/bin/sh -ex
 _scripts/install_packages.sh
+
+if [ "${SKIP_DOCKER}" = "true" ]; then
+  exit 0
+fi
+
 _scripts/install_docker.sh
