--- _scripts/oauth_mysql.sh
@@ -1,9 +1,0 @@
-#!/bin/bash -ex
-. ../../../_scripts/check_mysql.sh
-
-check_mysql
-mysqlStarted=$?
-
-if [ "$mysqlStarted" ]; then
-  node ./bin/server.js
-fi
