+++ _scripts/pushbox_db.sh
@@ -15,7 +15,7 @@ function on_singint() {
 trap on_singint INT
 
 docker run --rm --name pushbox_db \
-  --network fxa-net \
+  --network host \
   -p 4306:3306 \
   -e MYSQL_ROOT_PASSWORD=random \
   -e MYSQL_DATABASE=pushbox \
