+++ mysql_servers.json
@@ -83,7 +83,8 @@
     },
     {
       "name": "auth-server key server PORT 9000",
-      "script": "bin/key_server.js",
+      "script": "npm",
+      "args": ["start"],
       "cwd": "packages/fxa-auth-server",
       "env": {
         "DB": "mysql",
