--- packages/fxa-auth-server/scripts/start-server.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-node ./bin/key_server.js
-exit $PIPESTATUS[0]
