+++ packages/fxa-auth-server/scripts/test-local.sh
@@ -13,10 +13,6 @@ DEFAULT_ARGS="-R dot --recursive --timeout 5000 --exit"
 ./scripts/gen_vapid_keys.js
 node ./scripts/oauth_gen_keys.js
 
-if [ ! -e fxa-auth-db-mysql ] && [ -e "../../_scripts/clone-authdb.sh" ]; then
-  ../../_scripts/clone-authdb.sh
-fi
-
 GLOB=$*
 if [ -z "$GLOB" ]; then
   echo "Local tests"
