+++ packages/fxa-auth-server/test/test_server.js
@@ -8,7 +8,7 @@ const crypto = require('crypto');
 const EventEmitter = require('events');
 const P = require('../lib/promise');
 const mailbox = require('./mailbox');
-const createDBServer = require('../fxa-auth-db-mysql');
+const createDBServer = require('../../fxa-auth-db-mysql');
 const createAuthServer = require('../bin/key_server');
 const createMailHelper = require('./mail_helper');
 const createProfileHelper = require('./profile_helper');
