+++ packages/fxa-content-server/package.json
@@ -4,11 +4,10 @@
     "description": "Firefox Accounts Content Server",
     "scripts": {
         "build-production": "NODE_ENV=production grunt build",
-        "postinstall": "scripts/download_l10n.sh",
+        "postinstall": "cp server/config/local.json-dist server/config/local.json && scripts/download_l10n.sh",
         "lint": "eslint app server tests --cache",
         "lint:deps": "npm audit --json | audit-filter --nsp-config=.nsprc --audit=-",
         "start": "node scripts/check-local-config && grunt server",
-        "start-circle": "CONFIG_FILES=server/config/local.json,server/config/production.json,tests/ci/config_circleci.json node_modules/.bin/grunt build && CONFIG_FILES=server/config/local.json,server/config/production.json,tests/ci/config_circleci.json node_modules/.bin/grunt serverproc:dist",
         "start-production": "NODE_ENV=production grunt build && CONFIG_FILES=server/config/local.json,server/config/production.json grunt serverproc:dist",
         "start-remote": "scripts/run_remote_dev.sh",
         "test": "node tests/intern.js --unit=true",
