--- packages/fxa-content-server/tests/ci/config_circleci.json
@@ -1,16 +1,0 @@
-{
-    "csp": {
-        "enabled": false
-    },
-    "public_url": "http://127.0.0.1:3030",
-    "fxaccount_url": "https://fxaci.dev.lcip.org/auth",
-    "oauth_client_id": "98e6508e88680e1a",
-    "oauth_url": "https://oauth-fxaci.dev.lcip.org",
-    "profile_url": "https://fxaci.dev.lcip.org/profile",
-    "profile_images_url": "https://fxaci.dev.lcip.org/profile",
-    "allowed_parent_origins": ["https://123done-fxaci.dev.lcip.org"],
-    "use_https": false,
-    "i18n": {
-        "supportedLanguages": ["en"]
-    }
-}
