--- packages/fxa-content-server/tests/ci/config_verifier.json
@@ -1,9 +1,0 @@
-{
-    "logging": {
-        "loggers": {
-            "bid": {
-                "level": "ERROR"
-            }
-        }
-    }
-}
